refactor tests to use same codebase
